requirements.txt
accelerate==0.28.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.9.0
fonttools==4.51.0
fsspec==2024.3.1
gradio==4.25.0
gradio_client==0.15.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.4
importlib_resources==6.4.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
networkx==3.0
numpy==1.24.1
orjson==3.10.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
peft==0.9.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.1
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.28.1
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.5
safetensors==0.4.2
scipy==1.12.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
typer==0.12.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.13
uvicorn==0.29.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==11.0.3
yarg==0.1.9
